Editor: fix, improve and simplify getting context units
It reuses logic from `ViewRowResults` so we don't have to reinvent the wheel.
The endpoint has been greatly simplified too; on the one hand, the response
payload doesn't have redundant keys, and on the other hand it doesn't have to
take into account gaps anymore — it will always return a fixed set of units that
are before/after the unit id specified in the API endpoint.